Packaged the metrics and the data split
