+++ mlflow-credit-score/train.py
@@ -3,6 +3,7 @@ import warnings
 import sys
 import pandas as pd
 import numpy as np
+from creditscore.train import eval_metrics, data_split
 from sklearn.metrics import precision_score, recall_score, accuracy_score, f1_score
 from sklearn.model_selection import train_test_split
 from sklearn.ensemble import RandomForestClassifier
@@ -10,20 +11,14 @@ import mlflow
 import mlflow.sklearn
 
 import logging
+
 logging.basicConfig(level=logging.WARN)
 logger = logging.getLogger(__name__)
-
-def eval_metrics(actual, pred):
-    precision = precision_score(actual, pred, average='weighted')
-    recall = recall_score(actual, pred, average='weighted')
-    accuracy = accuracy_score(actual, pred)
-    f1 = f1_score(actual, pred, average='weighted')
-    return  precision, recall, accuracy, f1
 
 if __name__ == "__main__":
     warnings.filterwarnings("ignore")
     np.random.seed(40)
-    
+
     # Read the csv file
     csv_path = "../data/train.csv"
     try:
@@ -31,18 +26,13 @@ if __name__ == "__main__":
     except Exception as e:
         logger.exception(
             "Unable to download training & test CSV. Error: %s", e)
-        
-    data=data[["Interest_Rate", "Credit_Score"]]
+
+    data = data[["Interest_Rate", "Credit_Score"]]
     data_x = data.drop(["Credit_Score"], axis=1)
     data_y = data[["Credit_Score"]]
 
     # Split the data into training and test sets. (0.7, 0.3) split.
-    train_x, test_x, train_y, test_y = train_test_split(data_x, 
-                                                        data_y, 
-                                                        test_size=0.3, 
-                                                        stratify=data_y, 
-                                                        random_state=40, 
-                                                        shuffle=True)
+    train_x, test_x, train_y, test_y = data_split(data_x, data_y)
 
     n_estimators = int(sys.argv[1]) if len(sys.argv) > 1 else 100
     max_depth = int(sys.argv[2]) if len(sys.argv) > 2 else 3
